Fixed runtime compiler problem with unsigned int64 flags in object header.

+++ gldcore/rt/gridlabd.h
@@ -729,7 +729,7 @@ struct s_object_list {
 	unsigned int lock; /**< object lock */
 	unsigned int rng_state; /**< random number generator state */
 	TIMESTAMP heartbeat; /**< heartbeat call interval (in sim-seconds) */
-	uint64 flags; /**< object flags */
+	unsigned long long flags; /**< object flags */
 	/* IMPORTANT: flags must be last */
 }; /**< Object header structure */
